+++ sscanf_macro/src/attribute/struct.rs
@@ -22,8 +22,8 @@ impl FromAttribute<attr::Struct> for StructAttributeKind {
             }
             attr::Struct::Transparent => {
                 if let Some(value) = attr.value.as_ref() {
-                    let msg = format!("attribute `{}` does not take a value", attr.kind);
-                    return Error::err_spanned(value, msg); // checked in tests/fail/derive_struct_attributes.rs
+                    bail!(value => "attribute `{}` does not take a value", attr.kind);
+                    // checked in tests/fail/derive_struct_attributes.rs
                 }
                 Self::Transparent
             }