+++ tests/fail/missing_params.rs
@@ -4,5 +4,4 @@ fn main() {
     sscanf::sscanf!("5",);
     sscanf::sscanf!("5", 'x');
     sscanf::sscanf!("5", usize);
-    sscanf::sscanf_get_regex!();
 }